[web][refactor] move form payable form to another file.
